Add categories + category view
Add categories: Now you can add categories and while doing it choose which one note you want to assign. If you type a name of non-existent note - after creating a category a temporary note will be created which you can finish right after and it will automatically be assigned to the category you intended.

Category view: still working on the front-end part, but some basic design has been created. And it even manages to display photos of the notes in the selected category.